Fix misbehaving xinetd::service when $ensure = 'absent'
Previously the Puppet code tried to manage the $ensure and $enable parameters in
the service resource based on value of $::xinetd::ensure. That approach failed,
however, if the package/port had never been installed and its init script was
thus missing. Now the service resource is not managed at all if
$::xinetd::ensure is 'absent'.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>